# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * KidCall video conferencing API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Room {
  id: string;
  hostName: string;
  createdAt: string;
  active: boolean;
  hasPin: boolean;
}

export interface RoomInput {
  /** @minLength 1 */
  hostName: string;
  pin?: string;
}

export interface VerifyPinInput {
  pin: string;
}

export interface VerifyPinResponse {
  valid: boolean;
}

export interface Message {
  id: number;
  roomId: string;
  senderName: string;
  content: string;
  type: string;
  read: boolean;
  createdAt: string;
}

export interface MessageInput {
  /** @minLength 1 */
  senderName: string;
  /** @minLength 1 */
  content: string;
  type?: string;
}

export interface MarkReadResponse {
  updated: number;
}

export interface CallRequest {
  id: number;
  roomId: string;
  initiatedBy: string;
  status: string;
  createdAt: string;
}

export interface CallRequestResponse {
  callRequest: CallRequest | null;
}

export interface CallRequestInput {
  initiatedBy: string;
}

export interface UpdateCallRequestInput {
  status: string;
}

export interface ErrorResponse {
  error: string;
}