# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * KidCall video conferencing API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from 'zod';


/**
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  "status": zod.string()
})


/**
 * @summary Create a new call room
 */



export const CreateRoomBody = zod.object({
  "hostName": zod.string().min(1),
  "pin": zod.string().optional()
})


/**
 * @summary Get room details
 */
export const GetRoomParams = zod.object({
  "roomId": zod.coerce.string()
})

export const GetRoomResponse = zod.object({
  "id": zod.string(),
  "hostName": zod.string(),
  "createdAt": zod.string(),
  "active": zod.boolean(),
  "hasPin": zod.boolean()
})


/**
 * @summary Verify the room PIN before joining
 */
export const VerifyPinParams = zod.object({
  "roomId": zod.coerce.string()
})

export const VerifyPinBody = zod.object({
  "pin": zod.string()
})

export const VerifyPinResponse = zod.object({
  "valid": zod.boolean()
})


/**
 * @summary Get all messages for a room
 */
export const GetMessagesParams = zod.object({
  "roomId": zod.coerce.string()
})

export const GetMessagesResponseItem = zod.object({
  "id": zod.number(),
  "roomId": zod.string(),
  "senderName": zod.string(),
  "content": zod.string(),
  "type": zod.string(),
  "read": zod.boolean(),
  "createdAt": zod.string()
})
export const GetMessagesResponse = zod.array(GetMessagesResponseItem)


/**
 * @summary Send a message to the parent
 */
export const SendMessageParams = zod.object({
  "roomId": zod.coerce.string()
})



export const sendMessageBodyTypeDefault = `text`;

export const SendMessageBody = zod.object({
  "senderName": zod.string().min(1),
  "content": zod.string().min(1),
  "type": zod.string().default(sendMessageBodyTypeDefault)
})


/**
 * @summary Mark all messages in a room as read
 */
export const MarkMessagesReadParams = zod.object({
  "roomId": zod.coerce.string()
})

export const MarkMessagesReadResponse = zod.object({
  "updated": zod.number()
})


/**
 * @summary Get pending call request for a room
 */
export const GetCallRequestParams = zod.object({
  "roomId": zod.coerce.string()
})

export const GetCallRequestResponse = zod.object({
  "callRequest": zod.union([zod.object({
  "id": zod.number(),
  "roomId": zod.string(),
  "initiatedBy": zod.string(),
  "status": zod.string(),
  "createdAt": zod.string()
}),zod.null()]).nullable()
})


/**
 * @summary Initiate a call (ring the other side)
 */
export const CreateCallRequestParams = zod.object({
  "roomId": zod.coerce.string()
})

export const CreateCallRequestBody = zod.object({
  "initiatedBy": zod.string()
})


/**
 * @summary Answer or decline a call
 */
export const UpdateCallRequestParams = zod.object({
  "roomId": zod.coerce.string(),
  "callId": zod.coerce.number()
})

export const UpdateCallRequestBody = zod.object({
  "status": zod.string()
})

export const UpdateCallRequestResponse = zod.object({
  "id": zod.number(),
  "roomId": zod.string(),
  "initiatedBy": zod.string(),
  "status": zod.string(),
  "createdAt": zod.string()
})