# lib/api-zod/src/generated/types/callRequest.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * KidCall video conferencing API
 * OpenAPI spec version: 0.1.0
 */

export interface CallRequest {
  id: number;
  roomId: string;
  initiatedBy: string;
  status: string;
  createdAt: string;
}