# lib/api-zod/src/generated/types/callRequestInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * KidCall video conferencing API
 * OpenAPI spec version: 0.1.0
 */

export interface CallRequestInput {
  initiatedBy: string;
}