# lib/api-zod/src/generated/types/callRequestResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * KidCall video conferencing API
 * OpenAPI spec version: 0.1.0
 */
import type { CallRequest } from './callRequest';

export interface CallRequestResponse {
  callRequest: CallRequest | null;
}