# lib/api-zod/src/generated/types/room.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * KidCall video conferencing API
 * OpenAPI spec version: 0.1.0
 */

export interface Room {
  id: string;
  hostName: string;
  createdAt: string;
  active: boolean;
  hasPin: boolean;
}