# lib/api-zod/src/generated/types/roomInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * KidCall video conferencing API
 * OpenAPI spec version: 0.1.0
 */

export interface RoomInput {
  /** @minLength 1 */
  hostName: string;
  pin?: string;
}